requirements.txt
certifi==2025.1.31
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.3.0
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
Levenshtein==0.27.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.24.4
opencv-python-headless==4.7.0.68
packaging==24.2
pillow==11.1.0
pyclipper==1.3.0.post6
python-bidi==0.6.6
PyYAML==6.0.2
RapidFuzz==3.12.2
scikit-image==0.25.2
scipy==1.15.2
shapely==2.0.7
sympy==1.13.3
tifffile==2025.3.13
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
